fix loading zap, dat, frn, etc1. remove /Gm (min rebuild) 2. add _CRT_SECURE_NO_WARNINGS;_WIN32_WINNT=_WIN32_WINNT_MAXVER  (fort MGTools _WIN32_WINNT is not set, needs investigation) 3. add guards to GetContour() against Zdiv 4. add projects build order (Digit goes last) 5. fix INC_SEL_ broken line in EngLangD.rc 6. fix some double-to-int conversions with std::lround & static_cast<int>() 7. many conversion issues left 8. ElgLangD IntDir set to EngLangDTemp (was RusLangDTemp) 9. in MGTools:TIF_PRIN.CPP add %s to format:  _ftprintf(fd, _T("%sassoc-alpha"), sep) 10. TODO:  fix loading zap, dat, frn, etc1. remove /Gm (min rebuild) 2. add _CRT_SECURE_NO_WARNINGS;_WIN32_WINNT=_WIN32_WINNT_MAXVER  (fort MGTools _WIN32_WINNT is not set, needs investigation) 3. add guards to GetContour() against Zdiv 4. add projects build order (Digit goes last) 5. fix INC_SEL_ broken line in EngLangD.rc 6. fix some double-to-int conversions with std::lround & static_cast<int>() 7. many conversion issues left 8. ElgLangD IntDir set to EngLangDTemp (was RusLangDTemp) 9. in MGTools:TIF_PRIN.CPP add %s to format:  _ftprintf(fd, _T("%sassoc-alpha"), sep) 10. TODO:

+++ ImageTempl/ImageDoc.cpp
@@ -1026,7 +1026,7 @@ void CImageDoc::OnFileSaveAs()
     CString str;
     int nFilters=0;
 	
-	if(!pCtrls->GetCorrectFilterForSave(str, nFilters))
+	if(!pCtrls->GetCorrectFilterForSave(str, nFilters)) // modifies str, nFilters
 		return;
 
     CMainFrame* pMFr = GetMainFrame();
@@ -1048,7 +1048,7 @@ void CImageDoc::OnFileSaveAs()
     else
        fileDlg.m_ofn.nFilterIndex = 1;
 
-    if(fileDlg.m_ofn.nFilterIndex > nFilters)
+    if(fileDlg.m_ofn.nFilterIndex > static_cast<unsigned>(nFilters))
        fileDlg.m_ofn.nFilterIndex = 1;
 
     if(fileDlg.m_ofn.nFilterIndex)
